+++ scripts/init.mjs
@@ -1,11 +1,11 @@
 import * as helpers from "./helpers/index.mjs";
 import * as macros from "./macros/index.mjs";
-import { MODULE } from "./constants.mjs";
+import * as classes from './classes/index.mjs';
 import { registerSettings } from "./settings.mjs";
-const { MHLError, localize } = helpers;
 Hooks.on("init", () => {
   game.pf2emhl = {
     macros: {},
+    classes: {},
   };
   //helpers go in the root of the api object
   for (const [key, helper] of Object.entries(helpers)) {
@@ -15,10 +15,13 @@ Hooks.on("init", () => {
   for (const [key, macro] of Object.entries(macros)) {
     game.pf2emhl.macros[key] = macro;
   }
+  //classes also get subobjected
+  for (const [key, classObj] of Object.entries(classes)) {
+    game.pf2emhl.classes[key] = classObj;
+  }
   //shorthand when working at home
   if (game.modules.get("esheyw-transfer")?.active) {
     globalThis.mh = game.pf2emhl;
   }
-
   registerSettings();
 });