+++ scripts/macros/index.mjs
@@ -2,3 +2,4 @@ export * from './fascinatingPerformance.mjs';
 export * from './dropHeldTorch.mjs';
 export * from './lashingCurrents.mjs';
 export * from './recoverOldLashingCurrents.mjs';
+export * from './updateInitiativeSkills.mjs'