+++ CHANGELOG.md
@@ -10,7 +10,11 @@
 - fix imports
 
 ## Release 0.1.3
-- fix error string in oneTokenOnly()
+- fix error string in `oneTokenOnly()`
 - actually import targetHelpers
-- add some debug to applyOwnershipToFolderStructure()
-- fixed data path in setInitiativeStatistic()
+- add some debug to `applyOwnershipToFolderStructure()`
+- fixed data path in `setInitiativeStatistic()`
+
+## Release 0.1.4
+- add `updateInitiativeSkillsDialog()` macro
+- add `MHDialog` class